requirements.txt
opencv-python
numpy
imutils
easyocr
googletrans==4.0.0rc1
deep-translator
pillow
